refactor(platform): extract tmux availability cache to child module + gate sync (#4113)
Keeps platform/tmux.rs below the 1000-line giant threshold (1235->? after
verbatim move: cache consts/enum/struct/static/resolve + tests to
tmux/availability.rs); claude.rs freeze entry synced (-21 from
backend_routing extraction).
