+++ ARCHITECTURE.md
@@ -345,6 +345,8 @@ src/
 │   ├── automation_candidate_materializer/
 │   │   ├── allowed_path_tests.rs
 │   │   └── iteration_result_tests.rs
+│   ├── claude/
+│   │   └── backend_routing.rs
 │   ├── claude_e/
 │   │   ├── cancellation.rs
 │   │   ├── jsonl_parser.rs
@@ -888,6 +890,8 @@ src/
 │   │   ├── mod.rs
 │   │   └── provider.rs
 │   ├── platform/
+│   │   ├── tmux/
+│   │   │   └── availability.rs
 │   │   ├── binary_resolver.rs
 │   │   ├── dump_tool.rs
 │   │   ├── mod.rs
